requirements.txt
psutil
PyYAML
fastapi
uvicorn
pydantic
#transformers
#torch
cryptography
grpcio
grpcio-tools
protobuf
requests
dnspython
scapy
beautifulsoup4
stem